[disk][host]: move back Warnings from internal to disk and host.
fix #1377
